# docs/swagger.yaml
basePath: /api/v4
consumes:
- application/json
definitions:
  reaction.Rtype:
    enum:
    - LIKE
    - LIKE
    - LOVE
    - HAHA
    - WOW
    - SAD
    - ANGRY
    type: string
    x-enum-varnames:
    - DefaultRtype
    - RtypeLIKE
    - RtypeLOVE
    - RtypeHAHA
    - RtypeWOW
    - RtypeSAD
    - RtypeANGRY
...
  schemas.PostInputResponseDataSchema:
    properties:
      author:
        $ref: '#/definitions/schemas.UserDataSchema'
      comments_count:
        example: 35
        type: integer
      created_at:
        example: "2024-01-14T19:00:02.613124+01:00"
        type: string
      file_upload_data:
        $ref: '#/definitions/utils.SignatureFormat'
      image:
        example: https://img.url
        type: string
      reactions_count:
        example: 200
        type: integer
      slug:
        example: john-doe-d10dde64-a242-4ed0-bd75-4c759644b3a6
        type: string
      text:
        example: God Is Good
        type: string
      updated_at:
        example: "2024-01-14T19:00:02.613124+01:00"
        type: string
    type: object
  schemas.PostInputResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.PostInputResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.PostInputSchema:
    properties:
      file_type:
        example: image/jpeg
        type: string
      text:
        example: God is good
        type: string
    required:
    - text
    type: object
  schemas.PostResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.PostSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.PostSchema:
    properties:
      author:
        $ref: '#/definitions/schemas.UserDataSchema'
      comments_count:
        example: 35
        type: integer
      created_at:
        example: "2024-01-14T19:00:02.613124+01:00"
        type: string
      image:
        example: https://img.url
        type: string
      reactions_count:
        example: 200
        type: integer
      slug:
        example: john-doe-d10dde64-a242-4ed0-bd75-4c759644b3a6
        type: string
      text:
        example: God Is Good
        type: string
      updated_at:
        example: "2024-01-14T19:00:02.613124+01:00"
        type: string
    type: object
  schemas.PostsResponseDataSchema:
    properties:
      current_page:
        example: 1
        type: integer
      last_page:
        example: 100
        type: integer
      per_page:
        example: 100
        type: integer
      posts:
        items:
          $ref: '#/definitions/schemas.PostSchema'
        type: array
    type: object
  schemas.PostsResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.PostsResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.ReactionInputSchema:
    properties:
      rtype:
        allOf:
        - $ref: '#/definitions/reaction.Rtype'
        example: LIKE
    required:
    - rtype
    type: object
  schemas.ReactionResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.ReactionSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.ReactionSchema:
    properties:
      id:
        example: d10dde64-a242-4ed0-bd75-4c759644b3a6
        type: string
      rtype:
        example: LIKE
        type: string
      user:
        $ref: '#/definitions/schemas.UserDataSchema'
    type: object
  schemas.ReactionsResponseDataSchema:
    properties:
      current_page:
        example: 1
        type: integer
      last_page:
        example: 100
        type: integer
      per_page:
        example: 100
        type: integer
      reactions:
        items:
          $ref: '#/definitions/schemas.ReactionSchema'
        type: array
    type: object
  schemas.ReactionsResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.ReactionsResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.ResponseSchema:
    properties:
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.UserDataSchema:
    properties:
      avatar:
        example: https://img.url
        type: string
      name:
        example: John Doe
        type: string
      username:
        example: john-doe
        type: string
    type: object
...
  utils.SignatureFormat:
    properties:
      public_id:
        example: images/f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      signature:
        example: e1ba4683fbbf90b75ca22e9f8e545b18c6b24eae
        type: string
      timestamp:
        example: 1678828200
        type: integer
    type: object
info:
  contact: {}
  description: A Realtime Social Networking API built with Fiber
  title: SOCIALNET API
  version: "4.0"
paths:
...
  /feed/posts:
    get:
      description: This endpoint retrieves paginated responses of latest posts
      parameters:
      - default: 1
        description: Current Page
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.PostsResponseSchema'
      summary: Retrieve Latest Posts
      tags:
      - Feed
    post:
      description: This endpoint creates a new post
      parameters:
      - description: Post object
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/schemas.PostInputSchema'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.PostInputResponseSchema'
      security:
      - BearerAuth: []
      summary: Create Post
      tags:
      - Feed
  /feed/posts/{slug}:
    delete:
      description: This endpoint deletes a post
      parameters:
      - description: Post slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
      security:
      - BearerAuth: []
      summary: Delete a Post
      tags:
      - Feed
    get:
      description: This endpoint retrieves a single post
      parameters:
      - description: Post slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.PostResponseSchema'
      summary: Retrieve Single Post
      tags:
      - Feed
    put:
      description: This endpoint updates a post
      parameters:
      - description: Post slug
        in: path
        name: slug
        required: true
        type: string
      - description: Post object
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/schemas.PostInputSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.PostInputResponseSchema'
      security:
      - BearerAuth: []
      summary: Update Post
      tags:
      - Feed
  /feed/reactions/{focus}/{slug}:
    get:
      description: This endpoint retrieves paginated responses of reactions of post,
        comment, reply
      parameters:
      - description: 'Specify the usage. Use any of the three: POST, COMMENT, REPLY'
        in: path
        name: focus
        required: true
        type: string
      - description: Enter the slug of the post or comment or reply
        in: path
        name: slug
        required: true
        type: string
      - default: 1
        description: Current Page
        in: query
        name: page
        type: integer
      - description: 'Reaction Type. Must be any of these: LIKE, LOVE, HAHA, WOW,
          SAD, ANGRY'
        in: query
        name: reaction_type
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ReactionsResponseSchema'
      summary: Retrieve Latest Reactions of a Post, Comment, or Reply
      tags:
      - Feed
    post:
      description: This endpoint creates a new reaction.
      parameters:
      - description: 'Specify the usage. Use any of the three: POST, COMMENT, REPLY'
        in: path
        name: focus
        required: true
        type: string
      - description: Enter the slug of the post or comment or reply
        in: path
        name: slug
        required: true
        type: string
      - description: 'Reaction object. rtype should be any of these: LIKE, LOVE, HAHA,
          WOW, SAD, ANGRY'
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/schemas.ReactionInputSchema'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.ReactionResponseSchema'
      security:
      - BearerAuth: []
      summary: Create Reaction
      tags:
      - Feed
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer jwt_string' to correctly set the API Key
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"